Decouple the grafana manager and add text propagation option
Decouple the grafana manager  from the duplicator, and adding also a textpanel propagation script.
This script is useful for update the tabs in the homepages. (as we have 28 now).
