+++ src/python/CMSMonitoring/grafana_duplicator.py
@@ -7,7 +7,10 @@ import json
 import argparse
 import traceback
 import logging
+from grafana_manager import grafana_manager
+
 logging.basicConfig(level=os.environ.get("LOGLEVEL", "DEBUG"))
+
 
 class OptionParser:
     def __init__(self):
@@ -77,133 +80,6 @@ you can use several values using --datasource_replacement "cmsweb-k8s" "cmsweb-k
             default=False,
         )
 
-
-class grafana_manager:
-    def __init__(self, grafana_url=None, grafana_token=None, output_folder="./output"):
-        assert (
-            grafana_token and grafana_url
-        ), "both grafana_token and grafana_url are required."
-        self.GRAFANA_TOKEN = grafana_token
-        self.GRAFANA_URL = (
-            grafana_url[:-1] if grafana_url.endswith("/") else grafana_url
-        )
-        self.OUTPUT_FOLDER = output_folder
-
-    def dashboard_duplicator(
-        self,
-        dashboard_uid=None,
-        dashboards_query=None,
-        datasources_replacements=None,
-        title=None,
-        store_only=False,
-    ):
-        assert (
-            dashboard_uid or dashboards_query
-        ), "Either uid or query need to be specified"
-        replace_title = False
-        append_title = False
-        if title:
-            replace_title = True if dashboard_uid else False
-            append_title = not replace_title
-        dashboards = []
-        if dashboards_query:
-            dashboards = self.search_dasboards(
-                query=dashboards_query, uid=dashboard_uid
-            )
-        if dashboard_uid:
-            dashboards.append(self.get_dashboard(dashboard_uid))
-        # print(json.dumps(dashboards, indent=4))
-        replacements = (
-            dict(datasources_replacements) if datasources_replacements else {}
-        )
-        logging.debug('replacements {}', replacements)
-        for d in dashboards:
-            # Save a backup
-            self.__save("backup", d)
-            # Create the new dashboard
-            new_dashboard = {
-                "dashboard": d["dashboard"],
-                "overwrite": False,
-                "folderId": d["meta"]["folderId"],
-            }
-            new_dashboard["dashboard"]["id"] = None
-            new_dashboard["dashboard"]["uid"] = None
-            new_dashboard
-            new_title = (
-                title
-                if replace_title
-                else "{}_{}".format(new_dashboard["dashboard"]["title"], title)
-                if append_title
-                else new_dashboard["dashboard"]["title"]
-            )
-            new_dashboard["dashboard"]["title"] = new_title
-            self.__replace_dataset(new_dashboard, replacements)
-            self.__save("new", new_dashboard)
-            if not store_only:
-                logging.info("Try to create the dashboards in grafana")
-                logging.info(new_title)
-                logging.info(self.post_dashboard(new_dashboard))
-
-    def search_dasboards(self, query=None, uid=None):
-        params = {}
-        request_uri = None
-        request_uri = "{uri}/api/search".format(uri=self.GRAFANA_URL)
-        params["type"] = "dash_db"
-        params["query"] = query
-        dashboards_links = self.__make_request(request_uri, params=params)
-        return [self.get_dashboard(x["uid"]) for x in dashboards_links]
-
-    def get_dashboard(self, uid=None):
-        request_uri = "{uri}/api/dashboards/uid/{uid}".format(
-            uri=self.GRAFANA_URL, uid=uid
-        )
-        return self.__make_request(request_uri, params=None)
-
-    def post_dashboard(self, dashboard):
-        request_uri = "{uri}/api/dashboards/db".format(uri=self.GRAFANA_URL)
-        return self.__make_request(request_uri, params=dashboard, type="post")
-
-    def __make_request(self, uri, params, type="get"):
-        headers = {"Authorization": "Bearer {}".format(self.GRAFANA_TOKEN)}
-        verify = True
-        # The monit-grafana-dev uses an invalid certificate
-        if "-dev" in uri:
-            verify = False
-        response = None
-        if type == "get":
-            response = requests.get(uri, params=params, headers=headers, verify=verify)
-        elif type == "post":
-            response = requests.post(uri, json=params, headers=headers, verify=verify)
-        if response:
-            return json.loads(response.content)
-        else:
-            raise Exception("Something happen with the request, {}".format(response))
-
-    def __save(self, subfolder, d):
-        _folderId = (
-            d["meta"]["folderId"] if "meta" in d else d.get("folderId", "unknown")
-        )
-        filename = os.path.join(
-            self.OUTPUT_FOLDER,
-            subfolder,
-            "{folderId}_{title}.json".format(
-                folderId=_folderId, title=d["dashboard"].get("title", "unknown")
-            ),
-        )
-        os.makedirs(os.path.dirname(filename), exist_ok=True)
-        with open(filename, "w") as f:
-            json.dump(d, f, indent=4)
-
-    def __replace_dataset(self, element, replacements):
-        if isinstance(element, list):
-            for e in element:
-                self.__replace_dataset(e, replacements)
-        elif isinstance(element, dict):
-            for key in element:
-                if key == "datasource":
-                    element[key] = replacements.get(element[key], element[key])
-                elif isinstance(element[key], (dict, list)):
-                    self.__replace_dataset(element[key], replacements)
 
 def main():
     """
